requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
openpyxl==3.1.4
packaging==24.1
pillow==10.3.0
pypng==0.20220715.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
setuptools==70.0.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
